Jules was unable to complete the task in time. Please review the work done so far and provide feedback for Jules to continue.

+++ app/src/main/java/com/example/gestiontienda2/data/local/room/dao/ClientDao.kt
@@ -19,7 +19,7 @@ interface ClientDao {
     suspend fun deleteClient(client: ClientEntity)
 
     @Query("SELECT * FROM clients WHERE id = :clientId")
-    suspend fun getClientById(clientId: Long): ClientEntity?
+    suspend fun getClientById(clientId: Int): ClientEntity?
 
     @Query("SELECT * FROM clients")
     fun getAllClients(): Flow<List<ClientEntity>>
