+++ app/src/main/java/com/example/gestiontienda2/data/local/room/dao/OrderDao.kt
@@ -15,7 +15,7 @@ import kotlinx.coroutines.flow.Flow
 interface OrderDao {
 
     @Query("UPDATE orders SET status = :newStatus WHERE id = :orderId")
-    suspend fun updateOrderStatus(orderId: Long, newStatus: String)
+    suspend fun updateOrderStatus(orderId: Int, newStatus: String)
 
     @Insert(onConflict = OnConflictStrategy.REPLACE)
     suspend fun insertOrder(order: OrderEntity)
@@ -27,13 +27,13 @@ interface OrderDao {
     suspend fun deleteOrder(order: OrderEntity)
 
     @Query("SELECT * FROM orders WHERE id = :orderId")
-    suspend fun getOrderById(orderId: Long): OrderEntity?
+    suspend fun getOrderById(orderId: Int): OrderEntity?
 
     @Query("SELECT * FROM orders")
     fun getAllOrders(): Flow<List<OrderEntity>>
 
     @Query("DELETE FROM order_items WHERE orderId = :orderId")
-    suspend fun deleteOrderItemsForOrder(orderId: Long)
+    suspend fun deleteOrderItemsForOrder(orderId: Int)
 
     @Transaction
     @Query("SELECT * FROM orders")
@@ -41,5 +41,5 @@ interface OrderDao {
 
     @Transaction
     @Query("SELECT * FROM orders WHERE id = :orderId")
-    suspend fun getOrderWithItemsById(orderId: Long): OrderWithItems?
+    suspend fun getOrderWithItemsById(orderId: Int): OrderWithItems?
 }
